Rename as v1 and v2 in anticipation of upcoming Hercules release
