+++ whoc/interfaces/__init__.py
@@ -1,9 +1,9 @@
-from whoc.interfaces.hercules_actuator_disk_interface import (
+from whoc.interfaces.hercules_v1_interface import (
     HerculesADInterface,
     HerculesHybridADInterface,
 )
 from whoc.interfaces.hercules_battery_interface import HerculesBatteryInterface
-from whoc.interfaces.hercules_longrun_interface import (
+from whoc.interfaces.hercules_v2_interface import (
     HerculesHybridLongRunInterface,
     HerculesLongRunInterface,
 )
